+++ src/Common/EngineCore/Configuration/WorldManagementConfiguration.cs
@@ -23,4 +23,6 @@ namespace Sovereign.EngineCore.Configuration;
 public sealed class WorldManagementConfiguration : IWorldManagementConfiguration
 {
     public uint SegmentLength => 32;
+
+    public int SubscriptionRange => 3;
 }