+++ src/Common/EngineCore/Events/Details/WorldSegmentSubscriptionEventDetails.cs
@@ -1,0 +1,32 @@
+// Sovereign Engine
+// Copyright (c) 2023 opticfluorine
+// 
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+// 
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+// 
+// You should have received a copy of the GNU General Public License
+// along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
+using Sovereign.EngineCore.Components.Indexers;
+
+namespace Sovereign.EngineCore.Events.Details;
+
+public class WorldSegmentSubscriptionEventDetails : IEventDetails
+{
+    /// <summary>
+    ///     Entity ID of the subscribing player.
+    /// </summary>
+    public ulong EntityId;
+
+    /// <summary>
+    ///     Segment index of the subscribed world segment.
+    /// </summary>
+    public GridPosition SegmentIndex;
+}